requirements.txt
flask==2.0.1
tensorflow==2.8.0
numpy==1.22.3
pillow==9.0.1
pandas==1.4.2
scikit-learn==1.0.2
matplotlib==3.5.1
requests==2.27.1
